Move apiUmbrella functions into one file
